Add typing for RenderSlot and Teaser
